[ Init ]: Added configs to run the project, added repository unit tests and integration with service and resource
Changed password because i added the password as my mother name, and i forgot how to change that.
added repository unit tests to Usuario and Lançamento.
added integration test with service as UsuarioServiceImpl and resource of the User at "api/usuarios".
Added Gson dependency to convert to usarioDto to Json Object for resource